requirements.txt
pandas
numpy
python-okx
matplotlib
jupyter
Prophet
scikit-learn
statsmodels
pmdarima
xgboost
backtrader